feat(checker): Improve alone metrics error message
ErrUnexpectedAloneMetric is an error that returned from check when
targets with alone metrics set by user do not match actual targets with
alone metrics. The message of this error is not very clear for users.

Now in error description we place a hint which targets should be marked
as a targets with alone metrics and which should not.

Relates #428