feat(checker): Allow targets be single if it not declared
Targets in trigger can be marked as a "Single" what is mean that in
this target will be only one metric. If target have one metric and it is
not marked as "Single" check of this trigger will set trigger to
exception state as it was expected to have more than one metric in this
target.

This commit changes behavior of checker to allow targets not marked as
alone to have only one metric.

Relates #428