fix: Trigger simple check depends on tags
If trigger is simple we can enable realtime alerting for it. This means
that for each new metric we will check that trigger with new metric.

Simple trigger means that this trigger have only one metric. This is
able only if trigger have target without wildcards and braces.

With support for metrics with tags we need to add a new condition for
simple trigger% it should not use metrics with tags.

In this commit we rewrite simple trigger check.