feat(checker): Simplify ConvertForCheck function
In first versions of a new advanced mode after preparation metrics were
mixed. In one metrics set there was single and multi metrics targets.
Due to this on the step of conversion between
map[TargetName]map[MetricName]metrics to
map[MetricName]map[TargetName]metrics we had to check if some target is
alone metrics target and handle alone metrics targets separately. On the
other hand now on the step of preparation all targets are multi metric
targets and alone metric targets were filtered on the steps before so
we do not need to check alone metrics target on the step of conversion.

Relates #428