fix(notifier): Fixed race in notifier test
Race condition was when accessing to SelfCheckWorker.Heartbeats on read because creation of Heartbeats was in worker goroutine (Start -> selfStateChecker). Fixed race condition by moving heartbeats creation to constructor of SelfCheckWorker. Made heartbeats internal to pay attention that it is not public by design. Created error variables to indicate that self state monitor is not started.