fix: Added error handle. (#783)
If there is no data for metric, carbonapi consider it error `parser.ErrSeriesDoesNotExist`, but for Moira it is ok! It can happen when Graphite stopped or Moira clients send events instead of metrics.

There was wrapped error `parser.ErrSeriesDoesNotExist` that we don't handle, so some triggers gone in Exception state on cleanup of cli. Added handle.