Added new cmd to remove triggers (#790)
* feat(cli): Added new cmd

There is a problem - one of teams of Moira users have unused triggers. So it is a trash in database. Added command to remove triggers which have ID starting with prefix.

* refactor(cli): Renamed var

There was linter error and bug - shadow of global variable "triggers". Renamed global var for more detail because it has bigger scope that local var.