Mattermost support (#791)
* feat(notifier): Added Mattermost sender

Mattermost is an Open Source chat. Need to add Mattermost as delivery channel for Moira. Added new package for Sender that use mattermost-server dependency to send messages to Moira users via Bot.

* refactor(mock): Fix warnings of generate_mocks.sh

There were warnings:
```
#..\go\pkg\mod\github.com\golang\mock@v1.6.0\mockgen\mockgen.go:41:2: missing go.sum entry for module providing package golang.org/x/mod/modfile (imported by github.com/golang/mock/mockgen); to add:
#        go get github.com/golang/mock/mockgen@v1.6.0
#..\go\pkg\mod\github.com\golang\mock@v1.6.0\mockgen\mockgen.go:42:2: missing go.sum entry for module providing package golang.org/x/tools/imports (imported by github.com/golang/mock/mockgen); to add:
#        go get github.com/golang/mock/mockgen@v1.6.0
```

Use version for mockgen, it is recommended way to install mockgen. Remove `go get -u` because it has no sense, `github.com/golang/mock` is in go.mod.

* refactor: Use Gomock

Use generated mocks instead of manually created.

* test: Added manual integration test

There is a case when developer want to check integration with Mattermost locally. Added test for this case. There is build tag "manual" because this test need to run manually (not on common run, not on CI). Nice post about build tags - https://mickey.dev/posts/go-build-tags-testing/

* refactor(test): Name magic numbers

* feat: Add insecure TLS flag

Secure connection is important for production env, but is not for dev env.

* test: Tests with non-existent keys