fix: Fix response content-type
There was wrong content-type for response, it was 'text/plain'. Now it is JSON. Use render pkg functions to set StatusCode and JSON payload. It is because `responseWriter.Result()` use `snapHeader` field, which is 'snapshot of HeaderMap at first Write', so write operations to response need to use via render pkg functions, not manually.