fix: calculate RemoteAllowed based on remote/prom configs
Currently, RemoteAllowed controls if triggers of both "remote" types (remote graphite and prometheus) could be created, but it is calculated only based on remote graphite config.

In our case we only use prometheus triggers, so it is a little odd to fill in remote graphite config section to allow to create prometheus triggers.